Movies Collections Data Structure Change within Firebase
So far, each user had an object within their userdoc named "movies_collections", which contained arrays where the array-name was the collection name and array items were movies (by imdbID) added to that array/collection.

I've decided to go with a new structure where movies_collections is an object of objects, where each object within is a collection name, and the keys of each collection name object contains information such as createdBy, dateCreated, description, and of course, the movies array.